Move connection resources
