+++ workos/types/list_resource.py
@@ -25,7 +25,7 @@ from workos.types.directory_sync import (
 from workos.types.events import Event
 from workos.types.mfa import AuthenticationFactor
 from workos.types.organizations import Organization
-from workos.resources.sso import ConnectionWithDomains
+from workos.types.sso import ConnectionWithDomains
 from workos.resources.user_management import Invitation, OrganizationMembership, User
 from workos.types.workos_model import WorkOSModel
