+++ workos/types/sso/connection.py
@@ -1,4 +1,5 @@
-from typing import Literal
+from typing import Literal, Sequence
+from workos.types.sso.connection_domain import ConnectionDomain
 from workos.types.workos_model import WorkOSModel
 from workos.typing.literals import LiteralOrUntyped
 
@@ -53,3 +54,9 @@ class Connection(WorkOSModel):
     state: LiteralOrUntyped[ConnectionState]
     created_at: str
     updated_at: str
+
+
+class ConnectionWithDomains(Connection):
+    """Representation of a Connection Response as returned by WorkOS through the SSO feature."""
+
+    domains: Sequence[ConnectionDomain]
