+++ workos/types/sso/connection_domain.py
@@ -1,0 +1,8 @@
+from typing import Literal
+from workos.types.workos_model import WorkOSModel
+
+
+class ConnectionDomain(WorkOSModel):
+    object: Literal["connection_domain"]
+    id: str
+    domain: str
